Update LogHandler.php
Today if a log is handled successfully by Nightwatch it doesn't bubble the logs to further channels in the stack. 

In it's current state, Nightwatch will not allow logs to bubble to `webhook` because of it's position in the stack. This means that the order in a `.env`  _really_ matters.
```
LOG_STACK=single,nightwatch,webhook
```